requirements.txt

openai
openai-whisper
ffmpeg-python
srt
platformdirs
colorama
polib
tqdm
